Algebra.Field.AlgebraicReal

Description

Algebraic Real Numbers for exact computation

Since 0.4.0.0

Documentation

data Algebraic

Algebraic real numbers, which can be expressed as a root of a rational polynomial.

algebraic :: Unipol Rational -> Interval Rational -> Maybe Algebraic

Smart constructor. algebraic f i represents the unique root of rational polynomial f in the interval i. If no root is found, or more than one root belongs to the given interval, returns Nothing.

Operations

nthRoot :: Int -> Algebraic -> Maybe Algebraic

nthRoot n r tries to computes n-th root of the given algebraic real r. It returns Nothing if it's undefined.

See also nthRoot'.

nthRoot' :: Int -> Algebraic -> Algebraic

Unsafe version of nthRoot.

improve :: Algebraic -> Algebraic

improve r returns the same algebraic number, but with more tighter bounds.

approximate :: Rational -> Algebraic -> Rational

approximate eps r returns rational number r' close to r, with abs (r - r') < eps.

approxFractional :: Fractional r => Rational -> Algebraic -> r

Same as approximate, but returns Fractional value instead.

Equation solver

realRoots :: Unipol Rational -> [Algebraic]

realRoots f finds all real roots of the rational polynomial f.

complexRoots :: Unipol Rational -> [Complex Algebraic]

realRoots f finds all complex roots of the rational polynomial f.

CAUTION: This function currently comes with really naive implementation. Easy to explode.

Interval arithmetic

data Interval r

Constructors

Interval

Fields

lower :: !r

upper :: !r

representative :: (Additive r, Division r, Num r) => Interval r -> r

Choose representative element of the given interval.

includes :: Ord a => Interval a -> Interval a -> Bool

Test if the former interval includes the latter.

intersect :: (Monoidal a, Ord a) => Interval a -> Interval a -> Interval a

Takes intersection of two intervals.

Internal utility functions

strum :: Unipol Rational -> [Unipol Rational]

presultant :: (Euclidean k, CoeffRing k) => Unipol k -> Unipol k -> k

Pseudo resultant. should we expose this?

factors :: Unipol Rational -> [Unipol Rational]

realPartPoly :: Unipol Rational -> Unipol Rational

imagPartPoly :: Unipol Rational -> Unipol Rational

sqFreePart :: (Eq r, Euclidean r, Division r) => Unipol r -> Unipol r